pyproject.toml
[project]
name = "audiobook-gen"
version = "0.1.0"
description = ""
authors = [
    {name = "Lawrence Forman",email = "me@merklejerk.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.0.0)",
    "python-dotenv (>=0.19.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
